moving linear solver
